replaces queue with new impl
